🐛 Fix iscn format and iscn-dev functions not posting to correct endpoint
